Update carros.json with electric car data
This commit updates the carros.json file to include additional data for electric cars. The changes include adding the "tipo" (type) field to specify that the cars are electric, the "categoria" (category) field to indicate the type of car, the "tempoCarga" (charging time) field to specify the time required for charging, and the "custo" (cost) field to include the cost information for charging the car.